Add files via upload
Dataset Description
Overview
The dataset used in this project focuses on predicting restaurant ratings based on various features. It includes data points related to restaurant attributes such as location, cuisine, and cost. This dataset serves as the foundation for exploring predictive modeling and data analysis in the food and beverage industry.

Features
Restaurant Name: The name of the restaurant.
Location: The geographical location of the restaurant, often including city and state.
Cuisine: Type of cuisine offered by the restaurant (e.g., Italian, Chinese, Mexican).
Cost: Average cost of a meal or dining experience.
Rating: The restaurant's rating, which is the target variable for prediction.